+++ tests/testthat/test-edl_s3.R
@@ -1,0 +1,6 @@
+
+test_that("edl_s3", {
+
+  # tokens can only be used inside S3
+
+})