+++ tests/testthat/test-edl_set_token.R
@@ -4,4 +4,9 @@ test_that("edl_set_token", {
   token <- edl_set_token()
   expect_type(token, "character")
   expect_gt(nchar(token), 100)
+
+  url <- lpdacc_example_url()
+  r <- terra::rast(url, vsi=TRUE)
+  expect_true(inherits(r, "SpatRaster"))
+
 })