+++ NEWS.md
@@ -1,4 +1,12 @@
-# earthdatalogin
+# earthdatalogin 0.0.2
+
+* Adds support for netrc-based authentication. `edl_set_token()` works only
+  outside of `us-west-2`, `edl_s3_token()` works only inside `us-west-2`,
+  netrc works anywhere.  Additionally, netrc-based auth does not to be 
+  deactivated when accessing non-earthdatalogin URLs.  
+  
+  - `edl_netrc()` sets up netrc-based access.
+  - `edl_unset_netrc()` reverses this, but should rarely be needed.
 
 
 # earthdatalogin 0.0.1